Move code into src to try to fix lambda
