+++ .coveragerc
@@ -1,3 +1,3 @@
 [run]
-source = ds_caselaw_ingester/
+source = src/ds_caselaw_ingester/
 branch = True
