+++ scripts/test
@@ -1,1 +1,1 @@
-AWS_BUCKET_NAME=x PUBLIC_ASSET_BUCKET=y python -m pytest tests/* $*
+AWS_BUCKET_NAME=private-bucket PUBLIC_ASSET_BUCKET=public-bucket python -m pytest tests/* $*
